homework05: fix dependency injection
