+++ homework05/src/main/java/org/maslov/BookJdbcApp.java
@@ -2,7 +2,7 @@ package org.maslov;
 
 
 import org.maslov.model.Book;
-import org.maslov.repository.BookRepository;
+import org.maslov.repository.BookRepositoryImpl;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
@@ -14,7 +14,7 @@ import java.util.List;
 public class BookJdbcApp {
     public static void main(String[] args) {
         ApplicationContext context =SpringApplication.run(BookJdbcApp.class, args);
-        BookRepository br = context.getBean(BookRepository.class);
+        BookRepositoryImpl br = context.getBean(BookRepositoryImpl.class);
         List<Book> rv = br.findAll();
         for (Book b : rv) {
             System.out.println(b.getName());
